[CA] Updating controllers and routes
