+++ .env.example
@@ -1,0 +1,9 @@
+ENV=env
+PORT=port
+ACCOUNTNAME=account
+VTEX_API_KEY=apikey
+VTEX_API_TOKEN=apitoken
+KEY_UPLOAD=token
+TOKEN_SMS=token
+OPENCAGEDATA=token
+GOOGLEMAPSAPI=token
